v0.2: Se mejoro la interfaz de la clase Mapa agregando colores, un boton y etiquetas; se mejoro la interfaz de la Clase Interfaz_mapa, se le agrego una imagen de fondo, se corrigio varios errores en la graficacion del grafo y se modificaron las clases Edge y Punto_mapa. Por ultimo, se agrego la clase test donde se hizo un mapa de los distritos de arequipa, dicho mapa sera agregado al proyecto final, y se agrego la imagen del mapa de los distritos

+++ Edge.java
@@ -1,13 +1,14 @@
 
 public class Edge{
 	private Punto_mapa refDest;
-	private int weight;
+	private float weight;
 	private String label;
 	private String color;
+	
 	public Edge(Punto_mapa refDest) {
 		this(refDest,-1);
 	}
-	public Edge(Punto_mapa refDest, int weight) {
+	public Edge(Punto_mapa refDest, float weight) {
 		this.refDest= refDest;
 		this.weight= weight;
 	}
@@ -17,8 +18,8 @@ public class Edge{
 	public void setColor(String color) {this.color = color;}
 	public String getLabel() {return this.label;}
 	public void setLabel(String label) {this.label=label;}
-	public void setWeight(int weight) {this.weight=weight;}
-	public int getWeight() {return this.weight;}
+	public void setWeight(float weight) {this.weight=weight;}
+	public float getWeight() {return this.weight;}
 	public Punto_mapa opposite() {return this.refDest;}
 	
 	public boolean equals(Object o) {
